refacter: interview, persona 참조 관계 수정

+++ src/main/java/repit/repit_api_server/domain/userdata/interview/repository/PersonaRepository.java
@@ -3,5 +3,5 @@ package repit.repit_api_server.domain.userdata.interview.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import repit.repit_api_server.domain.userdata.interview.entity.Persona;
 
-public interface PersonaRepository extends JpaRepository<Persona, Integer> {
+public interface PersonaRepository extends JpaRepository<Persona, Long> {
 }
